# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: The Middle East Hustle API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles:
    get:
      operationId: listProfiles
      tags: [profiles]
      summary: List profiles
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: sector
          in: query
          schema:
            type: string
        - name: filter
          in: query
          schema:
            type: string
            enum: [featured, newest, most_viewed, most_followed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of profiles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileListResponse"
  /profiles/{id}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get a single profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileDetail"
        "404":
          description: Not found
...
components:
  schemas:
...
    PollOption:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        voteCount:
          type: integer
        percentage:
          type: number
      required:
        - id
        - text
        - voteCount
        - percentage
    Poll:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        context:
          type: string
        category:
          type: string
        categorySlug:
          type: string
        tags:
          type: array
          items:
            type: string
        pollType:
          type: string
          enum: [binary, multiple_choice, head_to_head, hot_take]
        options:
          type: array
          items:
            $ref: "#/components/schemas/PollOption"
        totalVotes:
          type: integer
        isFeatured:
          type: boolean
        isEditorsPick:
          type: boolean
        createdAt:
          type: string
        endsAt:
          type: string
          nullable: true
        relatedProfileIds:
          type: array
          items:
            type: integer
      required:
        - id
        - question
        - category
        - categorySlug
        - tags
        - pollType
        - options
        - totalVotes
        - isFeatured
        - isEditorsPick
        - createdAt
        - relatedProfileIds
...
    Profile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        headline:
          type: string
        role:
          type: string
        company:
          type: string
        sector:
          type: string
        country:
          type: string
        city:
          type: string
        imageUrl:
          type: string
          nullable: true
        isFeatured:
          type: boolean
        isVerified:
          type: boolean
        viewCount:
          type: integer
        associatedPollCount:
          type: integer
        quote:
          type: string
      required:
        - id
        - name
        - headline
        - role
        - sector
        - country
        - city
        - isFeatured
        - isVerified
        - viewCount
        - associatedPollCount
        - quote
    ProfileDetail:
      allOf:
        - $ref: "#/components/schemas/Profile"
        - type: object
          properties:
            summary:
              type: string
            story:
              type: string
            lessonsLearned:
              type: array
              items:
                type: string
            quote:
              type: string
            relatedPolls:
              type: array
              items:
                $ref: "#/components/schemas/Poll"
            similarProfiles:
              type: array
              items:
                $ref: "#/components/schemas/Profile"
          required:
            - summary
            - story
            - lessonsLearned
            - quote
            - relatedPolls
            - similarProfiles
    ProfileListResponse:
      type: object
      properties:
        profiles:
          type: array
          items:
            $ref: "#/components/schemas/Profile"
        total:
          type: integer
      required:
        - profiles
        - total
...