# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Middle East Hustle API
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  name: string;
  headline: string;
  role: string;
  company?: string;
  sector: string;
  country: string;
  city: string;
  imageUrl?: string | null;
  isFeatured: boolean;
  isVerified: boolean;
  viewCount: number;
  associatedPollCount: number;
  quote: string;
}